fix: renamed inversify due to clash with package
